requirements.txt

openai==0.27.8         # For DALL·E and other OpenAI API functionalities
pandas==2.1.1          # For handling Excel data
PyPDF2==3.0.1          # For reading and parsing PDFs
sqlalchemy==2.0.20     # For database connection
pillow==10.0.0         # For image handling and processing